Hard code gauss and pgauss
Use gauss as the main fitter, and always run pgauss.  This simplifies
the configuration, such that the weight is only used to configure pgauss
and meas_type is not longer accepted.  stamp sizes are now fixed because
gauss and pgauss used the same defaults.

Along the way I removed unused configs for the PSF fitting and LM
fitter.

The tests were adjusted.